% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaIPDA function.R
\name{metaIPDA}
\alias{metaIPDA}
\title{meta-analysis by IPDA}
\usage{
metaIPDA(
  dir.location,
  DV,
  DV.nature,
  IV,
  IV.nature,
  Moderator = "",
  Moderator.nature = "",
  Control = ""
)
}
\arguments{
\item{dir.location}{The directory location that contains individual studies data in csv format}

\item{DV}{the name of DV, must be in character format, e.g., "DV01"}
}
\value{
kable outputs of IPDA meta analysis results
}
\description{
add later
}
\examples{
output <- metaIPDA ()
}
